added base controller functionality

+++ core/src/ca/kanoa/zombieworld/Updateable.java
@@ -6,6 +6,6 @@ public interface Updateable {
      * Updates an object which supports it and gives the time in milliseconds since the last update
      * @param delta The time in milliseconds since the last update
      */
-    void update(double delta);
+    void update(long delta);
 
 }
